これで、OK

+++ boot/common.c
@@ -574,7 +574,7 @@ void text_gen(char *str, size_t max_size, const char *format, ...)
                     length = 16;
                 };
                 char buf[17];
-                UINT32 num = va_arg(args, UINT32);
+                UINT64 num = va_arg(args, UINT64);
                 hex_gen(buf, num, length);
                 const char *arg_str = buf;
                 while (*arg_str && dest < end)
